+++ CHANGES.md
@@ -8,6 +8,8 @@ next
 - Generate correct META files for sub-libraries (of the form `lib.foo`) that
   contain .js runtime files. (#3445, @hhugo)
 
+- Correctly infer targets for the `diff?` action. (#3457, fixes #2990, @greedy)
+
 2.5.1 (17/04/2020)
 ------------------